+++ components/kyma-operator/pkg/kymaoperation/steps/factory_test.go
@@ -1,4 +1,4 @@
-package actions
+package steps
 
 import (
 	"testing"
