+++ components/kyma-operator/pkg/kymaoperation/steps/step.go
@@ -1,4 +1,4 @@
-package actions
+package steps
 
 import (
 	"fmt"
