+++ components/kyma-operator/pkg/kymaoperation/steps/step_test.go
@@ -1,4 +1,4 @@
-package actions
+package steps
 
 import (
 	"errors"
