+++ components/kyma-operator/pkg/kymaoperation/steps/upgrade.go
@@ -1,4 +1,4 @@
-package actions
+package steps
 
 import (
 	"errors"
